Change Architecture to MVC and Create Models
