+++ src/app/controllers/instructors.js
@@ -1,20 +1,15 @@
-const { age, date } = require('../lib/utils');
+const { age, date, formatInstructors } = require('../lib/utils');
 const db = require('../config/db');
+const Instructor = require('../models/instructor');
 Intl = require('intl');
 
 module.exports = {
   index(req,res){
 
-    db.query(`SELECT * FROM instructors`, function(err,results){
-      if (err) return res.send('Database Error');
-
-      for (let row of results.rows){
-        row.services = row.services.split(',');
-      }
-
-
-      return res.render('instructors/index', {instructors: results.rows});
-    });
+    Instructor.all(function(instructors) {
+      formatInstructors(instructors);
+      return res.render('instructors/index', {instructors});
+    })
 
     
   },
@@ -30,37 +25,11 @@ module.exports = {
       }
     }
 
-    // const {avatar_url, name, services, gender } = req.body;
-    // const created_at = Date.now();
-    // let { birth } = req.body;
-    
-    const query = `
-      INSERT INTO instructors (
-        name,
-        avatar_url,
-        gender,
-        services,
-        birth,
-        created_at
-      ) VALUES ($1, $2, $3, $4, $5, $6)
-      RETURNING id
-    `
-    const values = [
-      req.body.name,
-      req.body.avatar_url,
-      req.body.gender,
-      req.body.services,
-      date(req.body.birth),
-      date(Date.now()),
-    ];
-
-    console.log(values);
-
-    db.query(query, values, function(err, results){
-      if (err) return res.send("Database Error");
-      
-      return res.redirect(`/instructor/${result.rows[0].id}`)
+    Instructor.create(req.body,function(instructor){
+      return res.redirect(`instructors/${instructor.id}`);
     });
+    
+    
     
     },
   show(req,res){
